+++ src/pages/profile/pages/my-flight/components/WaitingPaymentPage.tsx
@@ -1,0 +1,9 @@
+const WaitingPayment = () => {
+  return(
+    <section id="waitingPayment">
+      Waiting payment
+    </section>
+  )
+}
+
+export default WaitingPayment;
